+++ src/main/java/kitchenpos/dto/request/TableGroupCreateRequest.java
@@ -1,9 +1,6 @@
-package kitchenpos.dto;
+package kitchenpos.dto.request;
 
-import java.time.LocalDateTime;
 import java.util.List;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.TableGroup;
 
 public class TableGroupCreateRequest {
