+++ src/main/java/kitchenpos/ui/ProductRestController.java
@@ -2,8 +2,8 @@ package kitchenpos.ui;
 
 import java.util.List;
 import kitchenpos.application.ProductService;
-import kitchenpos.dto.ProductCreateRequest;
-import kitchenpos.dto.ProductResponse;
+import kitchenpos.dto.request.ProductCreateRequest;
+import kitchenpos.dto.response.ProductResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
